Add files via upload
Kaggle challenge - House Prices - Advanced Regression Techniques

Goal

It is your job to predict the sales price for each house. For each Id in the test set, you must predict the value of the SalePrice variable. 

Metric

Submissions are evaluated on Root-Mean-Squared-Error (RMSE) between the logarithm of the predicted value and the logarithm of the observed sales price. (Taking logs means that errors in predicting expensive houses and cheap houses will affect the result equally.)

Link: https://www.kaggle.com/competitions/house-prices-advanced-regression-techniques/overview/evaluation